Refactor `Surface` implementation into a dedicated module
Moved the `Surface` struct and related methods from `types.rs` to a new `surface.rs` module for better modularity and organization. Removed redundant methods and adjusted imports in affected files to align with the updated structure.
