+++ src/surfaces/mod.rs
@@ -6,5 +6,7 @@ mod types;
 mod visualization;
 mod utils;
 mod traits;
+mod surface;
 
-pub use types::{Surface, Point3D};
+pub use types::Point3D;
+pub use surface::Surface;
